Extract primitive related things into module
I’m not a massive fan of what this does to our module dependency
structure - this introduces cycles between the semantics and the prim
module. Still it’s nice to have things consolidated a bit.
