+++ fathom/src/core.rs
@@ -5,6 +5,7 @@ use crate::source::Span;
 use crate::StringId;
 
 pub mod binary;
+pub mod prim;
 pub mod semantics;
 
 /// Modules
